backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
langchain==0.1.0
langchain-core==0.1.22
langchain-openai==0.1.1
langchain-google-genai==0.0.11
openai==1.11.1
anthropic==0.7.7
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.6.1
httpx==0.26.0
python-multipart==0.0.9
tavily-python==0.3.0
html2text==2024.2.26
langgraph==0.1.1
copilotkit==0.1.20